curasense-ml/Mini-CDSS/requirements.txt
transformers
ipykernel
torch
crewai
crewai-tools
langchain
langchain-google-genai
langchain_community
tavily-python
python-dotenv
fastapi
bitsandbytes
accelerate
huggingface_hub